refactor: Corrigido enum

+++ src/product/dto/updateProductDTO.go
@@ -5,5 +5,5 @@ type UpdateProductDTO struct {
 	Size   string  `validate:"omitempty,size" json:"size"`
 	Value  float32 `validate:"omitempty,number,min=0" json:"value"`
 	Fabric string  `validate:"omitempty,min=2" json:"fabric"`
-	Stock  int64   `validate:"omitempty,number,min=0" json:"stock"`
+	Stock  uint32   `validate:"omitempty,number,min=0" json:"stock"`
 }
